lots of type fixes; types.h is now a convenience wrapper on top of the standard uintN_t routines
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@334 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/res/mem.h
@@ -19,7 +19,7 @@ enum
 	MEM_ZERO = 0x1000
 };
 
-extern void* mem_alloc(size_t size, uint align = 1, uint flags = 0, Handle* ph = 0);
+extern void* mem_alloc(size_t size, size_t align = 1, uint flags = 0, Handle* ph = 0);
 
 #define mem_free(p) mem_free_p((void*&)p)
 extern int mem_free_p(void*& p);
